polyMesh data-->Numpy

+++ pyFVM/Fluxes.py
@@ -50,25 +50,25 @@ class Fluxes():
         #值保存在面心上
         #face fluxes
         ## face flux linearization coefficients for cell C (cell of interest)
-        self.FluxCf=np.zeros((theNumberOfFaces))
+        self.FluxCf=np.zeros((theNumberOfFaces),dtype=float)
         ## face flux linearization coefficients for neighbouring cell
-        self.FluxFf=np.zeros((theNumberOfFaces))
+        self.FluxFf=np.zeros((theNumberOfFaces),dtype=float)
         ## non-linear face coefficients 
-        self.FluxVf=np.zeros((theNumberOfFaces))
+        self.FluxVf=np.zeros((theNumberOfFaces),dtype=float)
         ## total face flux (equal to FluxCf*phiC+FluxFf*phiF+FluxVf)
         # self.FluxTf={}
         # self.FluxTf= np.zeros((theNumberOfFaces),3)
-        self.FluxTf= np.zeros((theNumberOfFaces))
+        self.FluxTf= np.zeros((theNumberOfFaces),dtype=float)
 
         #值保存在体心上
         #Volume fluxes (treated as source terms)
-        self.FluxC=np.zeros((theNumberOfElements))
+        self.FluxC=np.zeros((theNumberOfElements),dtype=float)
         ## volume flux equal to source value times cell volume (Q_{C}^{phi} * Vc)
-        self.FluxV=np.zeros((theNumberOfElements))
+        self.FluxV=np.zeros((theNumberOfElements),dtype=float)
         # self.FluxT={}
-        self.FluxT= np.zeros((theNumberOfElements))
+        self.FluxT= np.zeros((theNumberOfElements),dtype=float)
         ## volume fluxes from previous time step
-        self.FluxC_old=np.zeros((theNumberOfElements))
+        self.FluxC_old=np.zeros((theNumberOfElements),dtype=float)
 
         # for root, directory,files in os.walk(Region.caseDirectoryPath + os.sep +str(Region.timeDirectory)):
         #     if not files:
